# API/docs/swagger.yaml
definitions:
  handlers.ConnectUserResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /users/connect:
    post:
      consumes:
      - application/json
      description: Authenticate user with provided email and password
      parameters:
      - description: User's email
        in: formData
        name: email
        required: true
        type: string
      - description: User's password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/handlers.ConnectUserResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Authenticate user
      tags:
      - users
swagger: "2.0"